refactor: chat router 경로 수정
